in middle match the slections bar must only contains 3 steps if it is in low You will be matched with positions that meet only some of your preferences and information of your CV (résumé). We will try to apply to as many jobs as we can, but we risk that some might not be totally suitable for your profile
if it is middle
You will be matched with positions that meet most of your preferences and information of your CV (résumé). We will try to apply to a lot of seemingly suitable jobs, but we risk that some might not be a 100% match with your profile.

if it is high
You will be matched with positions that meet all your preferences and information of your CV (résumé). We will take extra precautions to narrow down our search and apply only to jobs that match with the most of your specified details, to the degree possible.